crumb.py: an initial prototype Crumb build system implementation
Currently, it does not support partial rebuilds, and it doesn't
remember anything between run. Only goal at this stage is to see
that the hooks are working and we are able to schedule the tools
to continue running at the proper times, regardless of the point
of invocation.

Due to the design, it already supports parallelization of the
invoced tools.